Updated query and models to adhere hibernate 5 (#13)
* Updated query to retrieve admins with native sql query

* Updated query and models to adhere hibernate 5

+++ pdf-app/src/main/java/com/bitMiners/pdf/services/QuestionTypeService.java
@@ -1,16 +1,16 @@
 package com.bitMiners.pdf.services;
 
-import java.util.List;
-
 import com.bitMiners.pdf.domain.QuestionType;
+
+import java.util.List;
 
 public interface QuestionTypeService {
 
-	Integer saveQuestionType(QuestionType questionType);
-	
-	Integer updateQuestionType(QuestionType questionType);
-	
-	void deleteQuestionType(int id);
-	
-	List<QuestionType> getAllQuestionTypes();
+    Long saveQuestionType(QuestionType questionType);
+
+    Long updateQuestionType(QuestionType questionType);
+
+    void deleteQuestionType(Long id);
+
+    List<QuestionType> getAllQuestionTypes();
 }
